# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SenAI CRM Intelligence Platform",
    "description": "Autonomous CRM triage system with RAG, LangGraph agents, and live web intelligence.",
    "version": "1.0.0",
    "contact": {
      "name": "SenAI Platform Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/agent/dry-run/{email_id}": {
      "post": {
        "tags": ["Agent"],
        "summary": "Dry run agent without persisting",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dry run result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResult"
                }
              }
            }
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
...
    "/respond/{email_id}": {
      "post": {
        "tags": ["Actions"],
        "summary": "Send a manual reply",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reply_body", "user_id"],
                "properties": {
                  "reply_body": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reply sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/status/{job_id}": {
      "get": {
        "tags": ["Status"],
        "summary": "Get job status",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassificationResult": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string"
          },
          "sentiment": {
            "type": "string"
          },
          "sentiment_score": {
            "type": "number"
          },
          "urgency": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          },
          "requires_human": {
            "type": "boolean"
          },
          "escalation_reason": {
            "type": "string"
          },
          "suggested_reply": {
            "type": "string"
          },
          "detected_entities": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "reasoning": {
            "type": "string"
          }
        }
      },
      "AgentPlan": {
        "type": "object",
        "properties": {
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "action": {
                  "type": "string"
                },
                "thought": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "AgentRunResult": {
        "type": "object",
        "properties": {
          "email_id": {
            "type": "string"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationResult"
          },
          "plan": {
            "$ref": "#/components/schemas/AgentPlan"
          },
          "tool_results": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}